Improve unit test mocks by introducing more complex structures
= Motivation

This change is part of a bigger change that I'm planning to
upstream soon - the new and improved `member` command. In
order to test the command more thoroughly I needed structures
in our mocks that are a bit more complex than the existing
one. As the code change grew bigger, I decided to split it
into the testing infrastructure part and the actual command
part.

= Patch

This patch introduces adds a new array field in the test_struct
structure, and then introduces another struct that contains
test_struct pointers and an embedded struct. All of these cases
will be used by the tests added together with the new `member`
command.

= Testing

Besides ensuring that current tests don't fail I manually
checked the values of all the new mocks from the python
REPL like this:
```
$ python3 -i tests/__init__.py

>>> print(MOCK_PROGRAM['global_struct'])
(struct test_struct){
        .ts_int = (int)1,
        .ts_voidp = (void *)0xffffffffc0000000,
        .ts_array = (int [2]){ 252645135, -559038737, },
}

>>> print(hex(MOCK_PROGRAM['global_int'].address_of_().value_()))
0xffffffffc0000000
>>> print(hex(MOCK_PROGRAM['global_struct'].ts_array[0]))
0xf0f0f0f
>>> print(hex(drgn.cast('unsigned int', MOCK_PROGRAM['global_struct'].ts_array[1])))
0xdeadbeef

>>> print(MOCK_PROGRAM['global_cstruct'])
(struct complex_struct){
        .cs_structp = (struct test_struct *)0xffffffffc0a8aee0,
        .cs_struct = (struct test_struct){
                .ts_int = (int)10,
                .ts_voidp = (void *)0xffffffffd0000000,
                .ts_array = (int [2]){ -2147483648, 2147483647, },
        },
        .cs_structp_null = (struct test_struct *)0x0,
}

>>> print(hex(MOCK_PROGRAM['global_struct'].address_of_().value_()))
0xffffffffc0a8aee0
>>> print(hex(MOCK_PROGRAM['global_cstruct'].address_of_().value_()))
0xffffffffd0000000
>>> print(hex(MOCK_PROGRAM['global_cstruct'].cs_struct.ts_array[0]))
-0x80000000
>>> print(hex(MOCK_PROGRAM['global_cstruct'].cs_struct.ts_array[1]))
0x7fffffff
>>> print(hex(MOCK_PROGRAM['global_cstruct'].cs_structp_null))
0x0
```