Introduce count command
= Motivation

Sometimes we want to count the number of objects a pipeline is
producing. For example, in ZFS we may want to know how many
metaslabs are currently loaded. We'd currently do that with
the following pipeline that breaks out to the UNIX shell:
```
> spa | member spa_metaslabs_by_flushed | addr | walk | cast metaslab_t * | filter obj.ms_loaded == 1 | member ms_loaded ! wc -l
11
```
The above is not that bad but imagine if the question was how
many root slab caches exist in the system? Doing `slabs ! wc -l`
would be off by 2 because `wc` would count the header line and
the header separator. `slabs -H ! wc -l` would work but rarely
people remember the `-H` option outside of scripts. Finally,
users may do something awkward like `slabs | member name ! wc -l`
or `slabs | cast void * ! wc -l`. Given the situation, we can
avoid doing any of the above if SDB just had a command that
can count the objects passed to it through the pipeline.

= Patch

This commit introduces `count` which does exactly that. The
above examples with this command change respective to:
```
> spa | member spa_metaslabs_by_flushed | addr | walk | cast metaslab_t * | filter obj.ms_loaded == 1 | count
(unsigned long long)9
```
and
```
> slabs | count
(unsigned long long)132
```

There is no reason currently why the command needs to return
an object (an unsigned long long to be specific). That said,
it would be nice to have in the future where SDB could have
a subshell syntax construct where an object would need to be
returned.

= Testing/Verification

A few unit tests are added by the patch. I did some manual
testing while poking around on a system while trying to develop
other SDB commands.