Fix `spa <poolname>` functionality
= Background

As part of the port from crash-python to drgn there was a bug
filed (issue #12) pointing out that `spa <poolname` stopped
working.
Example output:
```
> spa
ADDR           NAME
------------------------------------------------------------
0xffff959e94210000 rpool

> spa rpool
ADDR           NAME
------------------------------------------------------------
>
```

Another issue that's visible from this is that the header
of the second columnt ("NAME") is not aligned correctly.

= Patch

This pass fixes both issues and the command now looks like
this:
```
> spa
ADDR               NAME
------------------------------------------------------------
0xffff9d0adbe2c000 application
0xffff9d0a2dd28000 prod-data
0xffff9d0ae5040000 rpool
0xffff9d0a2bdb0000 support

> spa rpool
ADDR               NAME
------------------------------------------------------------
0xffff9d0ae5040000 rpool

> spa rpool | member spa_name
(char [256])"rpool"
```

= Testing

The only testing that I did was the manual testing above.

 = Github Issue Tracker Automation

Closes #12