Leverage Github Actions for automated testing
This change adds the necessary files and logic to take advantage of
GitHub Actions for running our automated tests. Currently we rely on
TravisCI to do this, but we'd like to transition over to GitHub Actions,
for the following reasons:

  1. The Actions UI is more integrated with GitHub PRs
  2. Best practices at Delphix suggests using Actions

While this change is adding support for using GitHub Actions, it's not
currently disabiling the use of TravisCI. Since I don't have much
experience with Actions, I'd like to use both Actions and TravisCI
simultaneously until we have sufficient confidence in Actions; at which
point we can disable the TravisCI tests, and rely solely on Actions.