Add a help command
Adds 'help' to print out the description of either the given command or
all available commands. The text for the message is taken from the class
documentation (i.e. cls.__doc__). By default, the command prints a help
summary, but the optional -v parameter prints the entire text of the
help message. This implies class documentation strings should follow a
format similar to commit messages:
    one line description of the command
    optional, possibly multi-line description of the command