Handling Command Failures - Part 1
= Summary

Do not break out of a session when a command crashes or a
bug shows up in the SDB infrastructure. Prompt the user
to file a bug and keep the session instead. State is not
retained between commands issued thus this is safe to do
for now and provides a more user-friendly experience.

= Testing

```
sdb> echo 1234 | cast dnode_t * | pp
sdb encountered an internal error due to a bug. Here's the
information you need to file the bug:
----------------------------------------------------------
Target Info:
	ProgramFlags.IS_LIVE|IS_LINUX_KERNEL
	Platform(<Architecture.X86_64: 1>, <PlatformFlags.IS_LITTLE_ENDIAN|IS_64_BIT: 3>)

Traceback (most recent call last):
  File "/usr/local/lib/python3.6/dist-packages/sdb-0.1.0-py3.6.egg/sdb/internal/repl.py", line 88, in eval_cmd
    for obj in objs:
  File "/usr/local/lib/python3.6/dist-packages/sdb-0.1.0-py3.6.egg/sdb/__init__.py", line 182, in invoke
    execute_pipeline_term(prog, first_input, pipeline)
  File "/usr/local/lib/python3.6/dist-packages/sdb-0.1.0-py3.6.egg/sdb/__init__.py", line 103, in execute_pipeline_term
    pipeline[-1].call(this_input)
  File "/usr/local/lib/python3.6/dist-packages/sdb-0.1.0-py3.6.egg/sdb/commands/pretty_print.py", line 48, in call
    self.names, i.type_))
TypeError: command "['pretty_print', 'pp']" does not handle input of type dnode_t *
----------------------------------------------------------
Link: https://github.com/delphix/sdb/issues/new
```

= Github Issue Tracker Automation

Closes #87