Run Actions on a regular basis
Currently our tests are not entirely reproducible, since we don't pin
our dependencies to specific versions. Thus, frequently we'll open a
pull request, only to realize that a depdency has changed, and a new
test regression was silently introduced.

To help us notice these regressions sooner, this change is augmenting
our tests to run on a daily basis. This way, we'll continually test with
new versions of our dependencies, which should help us detect when
regressions are introduced, and allow us to fix the issue before the
test failure cascades into an unrelated pull request.