Fix commands that define non-generator call()
Dealing with functions that may or may not be generators is tricky.
The call() method for subclasses of Command is one example.
Unfortunately, #100 broke non-generator Commands like `pp` and `help`.
The specific problem is twofold:

1. changing execute_pipeline_term() to call a generator
(massage_input_and_call), which doesn't work (you have to `yield from`
or iterate over a generator).  The solution is to iterate this
generator, even when we expect it not yield anything.

2. massage_input_and_call() `yield from` a non-generator function
(self.call).  The solution is to only `yield from` `self.call` if it
returns something.