Implement _call_one() method for Commands Handle all uncaught invalid memory accesses in the pipeline Migrate positive tests cases of filter and member to integration tests Fix new pylint error for input_type
Note #1:
Test cases from filter and member had to be migrated because some
of our mocks were not backed by valid memory and the pipeline can
now detect that, that's why this change was bundled together. I
decided to not let go of the negative tests yet though as they
are still useful in checking that the correct programmatic
exception is thrown.

Note #2:
The error messages introduced when dereferencing invalid memory
are not optimal, partly because drgn doesn't give us a lot of
flexibility. The plan is to re-write them properly once we have
the relevant infrastructure ready underneath us.