Introduce ptype command
Note that the argument parsing could be made smarter (and more
complex) to allow the user to use C keywords like "struct"
before the type. We just leave it to SDB to figure out by
itself for now until we figure out how we want to deal with this
in this and other commands in the repo.