Disable python bindings for libkdumpfile temporarily
Due to issue #160 introduced by recent changes in libkdumpfile
our nightly and PR automation have been failing because the
library can't compile. Until the issue is resolved in the library
itself or on our side (by creating a proper environment for the
github action) we disable the compilation of the Python bindings
of the library. The impact is almost non-existent as these
bindings are not used in any way by SDB nor DRGN.