dbuf command filtering does not work in all invocations (#184)
The dbuf command can select dbufs that match certain parameters (e.g.
'-l 1' for level-1 dbufs).  This works when `dbuf` is the first command
in the pipeline, and when it is used as a locator with a dnode_t* as
input.  But it doesn't work in other circumstances, for example
when it's being used only as a pretty-printer (with the list of dbufs
piped in), or when it's used in the middle of a pipeline.

This change adds the filter calls to this use cases, and adds
corresponding test cases