Fix latest pylint errors
Pylint started being more strict about the use of Python's
raise-from syntax explained here:
https://www.python.org/dev/peps/pep-3134/

The main idea is to provide a proper traceback whenever an
exception is caught and rethrown. Specifically in our case
though the use of raise-from shouldn't make any difference
for us because SDB is a REPL and prints the error messages
of our sdb-defined exceptions (which are basically faulty
exits of SDB commands).