ptype: remove anonymous union test and improve error handling
In the past we used to assume lazy type rules from drgn and printing
an anonymous union or struct from its typedef just worked.
Unfortunately, with the latest updates for types in drgn, this no
longer works so part of this commit removes the expectation of
this functionality from our tests. We may want to introduce this
functionality again in the future implemented in a different way.
Fortunately, these kind of types are rare.

The second part of this commit improves the error handling of
looking up types.