Replace current PR template on Github
= Problem
The default PR template that we inherit from the Delphix Github
organization is too generic and some of its fields are not applicable
to SDB.

= Solution
This patch introduces a new template that is more suitable for our
needs.

= Notes To Reviewers
The choice of using `=` instead of `#` (Markdown Header) for
the template sections is for the case that the submitter decides
to copy paste the template in their commit description in git.
The `#` character is used for comments in git descriptions so
these section headers would disappear once the commit description
is written.