DLPX-84407 Refactor slub wrappers on SDB to use new drgn helpers
= Problem

While working on the ZIO command for ZFS, I found a bug in the
implementation of our SLUB cache walker. The bug was that we'd allow the
walker to return objects that were actually free and sometimes would
even dereference internal SLUB cache freelist metadata that would return
the message "inconsistent freelist address XXX".

Since this wrappers were first written, the same logic was also
incorporated to drgn in the form of helpers. Due to another bug in those
helpers though, we never attempted to use them.

= Solution

Fortunately the bug is now fixed:
https://github.com/osandov/drgn/commit/2bb0f9861c7ff4887baa5d5c72f6d5277005835a

And we can go ahead and use those helpers, no longer having to maintain
our own.

= This Patch

Besides replacing our own code with the new helper code we regenerate
the right regression output for our reference crash dump in the test
suite.