DLPX-87560 sdb: dmesg broken for 5.15 kernels
This patch replaces our hand-written implementation of dmesg
which doesn't work for most vanilla kernel versions after 5.10
with the drgn helper which stays up to date more often.

In order for this to work I had to change the type of our
dmesg command from a Locator+PrettyPrinter to a standard
command. This is not a big deal as the only reason we would
ever walk and filter dmesg would be to filter the kernel logs
by level. I've added this as an argument to the command.

= Github Issue Tracker Automation

Closes #336